+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### New features
+
+- Implemented `Hash` for `Code`.
+- Implemented `serde::Serialize` and `serde::Deserialize` for `Code`, behind new `serde` feature
+  flag.
+
 ## 0.1.0
 
 Initial release.